Improve Lyric Card Search (#16)
* Add page_data method to api.GeniusT

* Add utils.fix_section_header and SECTION_HEADERS regex expression

* Use utils.fix_section_header in song.display_lyrics

* Improve matching lyrics in lyric_card
Compare the similarity of the input and the lines of the snippet using the Levenshtein algorithm. And then compare the found lines with the lines in the full lyrics. This is much easier than trying to match the lyrics character by character

* Improve matching lyrics in inline query lyric_card

* Remove extra lyrics image in inline query
The solution doesn't seem clean and straightforward to me and it's also yet another API call which in this case is an upload and we can avoid.

* isorted

* Add python-levenshtein to requirements

Co-authored-by: allerter <8884518-allerter@users.noreply.gitlab.com>